Implement default layout
This is a primary implementation of the default layout component. In order
to complete it, a footer will be necessary.

Why did I not complete the layout design, then? Because the layout it's
not ready, and I'm doing it in such a way the layout becomes pleasant to
the eyes. I'm not rushing design elements implementation.

Next step is to build a new route for user operations.